Split the button and counter into three files
