+++ dom-merge-conflict/tasks/buttons-and-counter/src/app.js
@@ -1,3 +1,6 @@
+import { Header } from "./header";
+import { Main } from "./main";
+
 //increments the number in a node's text
 function increment(node) {
   let current = node.textContent;
@@ -7,18 +10,10 @@ function increment(node) {
 export function App() {
   const body = document.createElement("body");
 
-  const header = document.createElement("header");
-  header.innerHTML = `
-        <h1>Number Counter</h1>
-        <p>A simple counter. Press increment to increase the count by one.</p>
-    `;
+  const header = Header();
   body.appendChild(header);
 
-  const main = document.createElement("main");
-  main.innerHTML = `
-        <p id="counter" data-testid="counter">0</p>
-        <button id="increment">Increment</button>
-    `;
+  const main = Main();
   body.appendChild(main);
 
   const button = body.querySelector("#increment");
