+++ dom-merge-conflict/tasks/buttons-and-counter/src/main.js
@@ -1,0 +1,9 @@
+export function Main() {
+    const main = document.createElement("main");
+    main.innerHTML = `
+        <p id="counter" data-testid="counter">0</p>
+        <button id="increment">Increment</button>
+    `;
+
+    return main;
+}
