+++ package.json
@@ -1,1 +1,8 @@
-{}
+{
+  "scripts": {
+    "test": "jest"
+  },
+  "devDependencies": {
+    "jest": "^30.3.0"
+  }
+}
